fix: proxy /t/ in worker so posthog trailing slash isn't 301'd
